Refactor scaling-opengauss module into pipeline-opengauss module
