+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-dialect/shardingsphere-data-pipeline-opengauss/src/main/resources/META-INF/services/org.apache.shardingsphere.scaling.core.spi.ScalingEntry
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.scaling.opengauss.OpenGaussScalingEntry
+org.apache.shardingsphere.data.pipeline.opengauss.OpenGaussScalingEntry
